refactor(infra): renames <NotFound/> component-page folder
